requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
crypto==1.4.1
cryptography==36.0.2
cycler==0.11.0
Flask==2.1.1
Flask-Cors==3.0.10
fonttools==4.32.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.0
Naked==0.1.31
numpy==1.22.3
opencv-python==4.3.0.36
packaging==21.3
Pillow==9.1.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
rsa==4.8
shellescape==3.8.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
